requirements.txt
Django==1.8.4
Pillow==2.9.0
django-cors-headers==1.1.0
django-filter==0.11.0
django-grappelli==2.7.1
django-haystack==2.4.0
django-model-utils==2.3.1
djangorestframework==3.2.3
drf-haystack==1.5.2
factory-boy==2.5.2
psycopg2==2.6.1
